Configure datetime on wich reset environment

+++ app/Jobs/ResetEnvironment.php
@@ -55,42 +55,51 @@ class ResetEnvironment implements ShouldQueue
 
     public function handle(): void
     {
-        logger()->info(
-            'Resetting "' . config('app.env') . '" environment',
-            [
-                'job' => JobType::RESET_ENVIRONMENT,
-            ],
-        );
+        if (!app()->environment('production')) {
+            logger()->info(
+                'Resetting "' . config('app.env') . '" environment',
+                [
+                    'job' => JobType::RESET_ENVIRONMENT,
+                ],
+            );
 
-        Redis::connection(env('CACHE_CONNECTION'))->client()->flushdb();
-        Redis::connection(env('SESSION_CONNECTION'))->client()->flushdb();
-        Redis::connection(env('QUEUE_REDIS_CONNECTION'))->client()->flushdb();
+            Redis::connection(env('CACHE_CONNECTION'))->client()->flushdb();
+            Redis::connection(env('SESSION_CONNECTION'))->client()->flushdb();
+            Redis::connection(env('QUEUE_REDIS_CONNECTION'))->client()->flushdb();
 
-        $results = collect($this->commands)->mapToGroups(function ($command) {
-            try {
-                Artisan::call($command['command'], $command['parameters'] ?? []);
+            $results = collect($this->commands)->mapToGroups(function ($command) {
+                try {
+                    Artisan::call($command['command'], $command['parameters'] ?? []);
 
-                return [
-                    'completed' => [
-                        'command' => $command['command'],
-                    ],
-                ];
-            } catch (Exception $exception) {
-                report($exception);
-                $message = $exception->getMessage();
+                    return [
+                        'completed' => [
+                            'command' => $command['command'],
+                        ],
+                    ];
+                } catch (Exception $exception) {
+                    report($exception);
+                    $message = $exception->getMessage();
 
-                return [
-                    'failed' => [
-                        'command' => $command['command'],
-                        'message' => $message,
-                    ],
-                ];
-            }
-        });
+                    return [
+                        'failed' => [
+                            'command' => $command['command'],
+                            'message' => $message,
+                        ],
+                    ];
+                }
+            });
 
-        event(new EnvironmentResetCompleted(
-            empty($results->get('completed')) ? [] : $results->get('completed')->all(),
-            empty($results->get('failed')) ? [] : $results->get('failed')->all(),
-        ));
+            event(new EnvironmentResetCompleted(
+                empty($results->get('completed')) ? [] : $results->get('completed')->all(),
+                empty($results->get('failed')) ? [] : $results->get('failed')->all(),
+            ));
+        } else {
+            logger()->info(
+                'Not resetting "' . config('app.env') . '" environment',
+                [
+                    'job' => JobType::RESET_ENVIRONMENT,
+                ],
+            );
+        }
     }
 }
